Time: 0 ms (100.00%) | Memory: 63 MB (48.50%) - LeetSync

+++ 11-container-with-most-water/container-with-most-water.cpp
@@ -1,26 +1,26 @@
 class Solution {
 public:
     int maxArea(vector<int>& height) {
-
-        int left = 0;
+        int left =0;
         int right = height.size()-1;
-        long long area = 0;
+        int area = 0;
 
-        while(left<right){
-            long long width = right - left;
-            long long h = min(height[left], height[right]);
+        while(left < right){
+
+            int width = right-left;
+            int h = min(height[left], height[right]);
+
             area = max(area, width*h);
 
-            if(height[left]<height[right]){
+            if(height[left] < height[right]){
                 left++;
             }
-
             else{
                 right--;
             }
+
         }
 
         return area;
-        
     }
 };
